make a new file for environment.ts
